+++ SmartWalletApp/Features/Home/Data/Models/Responses/FinancialGoalSummaryResponse.swift
@@ -4,5 +4,5 @@ struct FinancialGoalSummaryResponse: Decodable {
     let totalGoalCount: Int
     let totalSavedAmount: Decimal
     let totalTargetAmount: Decimal
-    let completionPercentage: Int
+    let completionPercentage: Decimal
 }
